+++ django_mysql/forms.py
@@ -5,6 +5,7 @@ from __future__ import (
 
 import json
 
+import django
 from django import forms
 from django.core import validators
 from django.core.exceptions import ValidationError
@@ -225,26 +226,49 @@ class SimpleSetField(forms.CharField):
             raise ValidationError(errors)
 
 
+class InvalidJSONInput(six.text_type):
+    pass
+
+
+class JSONString(six.text_type):
+    pass
+
+
 class JSONField(forms.CharField):
     default_error_messages = {
         'invalid': _("'%(value)s' value must be valid JSON."),
     }
-
-    def __init__(self, **kwargs):
-        kwargs.setdefault('widget', forms.Textarea)
-        super(JSONField, self).__init__(**kwargs)
+    widget = forms.Textarea
 
     def to_python(self, value):
+        if django.VERSION[:2] >= (1, 9) and self.disabled:
+            return value
         if value in self.empty_values:
             return None
+        elif isinstance(value, (list, dict, int, float, JSONString)):
+            return value
         try:
-            return json.loads(value)
+            converted = json.loads(value)
         except ValueError:
             raise forms.ValidationError(
                 self.error_messages['invalid'],
                 code='invalid',
                 params={'value': value},
             )
+        if isinstance(converted, six.text_type):
+            return JSONString(converted)
+        else:
+            return converted
+
+    def bound_data(self, data, initial):
+        if django.VERSION[:2] >= (1, 9) and self.disabled:
+            return initial
+        try:
+            return json.loads(data)
+        except ValueError:
+            return InvalidJSONInput(data)
 
     def prepare_value(self, value):
+        if isinstance(value, InvalidJSONInput):
+            return value
         return json.dumps(value)