+++ django_mysql/models/functions.py
@@ -3,6 +3,8 @@ from __future__ import (
     absolute_import, division, print_function, unicode_literals
 )
 
+import json
+
 from django.db import DEFAULT_DB_ALIAS, connections
 from django.db.models import Field as DjangoField
 from django.db.models import CharField, Func, IntegerField, TextField, Value
@@ -260,6 +262,15 @@ class JSONLength(Func):
         super(JSONLength, self).__init__(*exprs, output_field=output_field)
 
 
+class JSONValue(Func):
+    function = 'CAST'
+    template = '%(function)s(%(expressions)s AS JSON)'
+
+    def __init__(self, expression):
+        json_string = json.dumps(expression, allow_nan=False)
+        super(JSONValue, self).__init__(Value(json_string))
+
+
 class BaseJSONModifyFunc(Func):
     def __init__(self, expression, data):
         from django_mysql.models.fields import JSONField
@@ -276,7 +287,7 @@ class BaseJSONModifyFunc(Func):
             exprs.append(path)
 
             if not hasattr(value, 'resolve_expression'):
-                value = Value(value)
+                value = JSONValue(value)
 
             exprs.append(value)