+++ django_mysql/monkey_patches.py
@@ -7,29 +7,14 @@ import functools
 
 from django.conf import settings
 from django.db.backends.mysql.base import CursorWrapper
-from django.utils.functional import cached_property
 
 from django_mysql.rewrite_query import REWRITE_MARKER, rewrite_query
-from django_mysql.utils import connection_is_mariadb
 
 
 def patch():
-    # Fine to patch straight on since it's a cached_property descriptor
-    patch_ConnectionWrapper_is_mariadb()
-
     # Depends on setting
     if getattr(settings, 'DJANGO_MYSQL_REWRITE_QUERIES', False):
         patch_CursorWrapper_execute()
-
-
-def patch_ConnectionWrapper_is_mariadb():
-    from django.db.backends.mysql.base import DatabaseWrapper
-
-    @cached_property
-    def is_mariadb(self):
-        return connection_is_mariadb(self)
-
-    DatabaseWrapper.is_mariadb = is_mariadb
 
 
 def patch_CursorWrapper_execute():