+++ tests/testapp/test_forms.py
@@ -3,10 +3,14 @@ from __future__ import (
     absolute_import, division, print_function, unicode_literals
 )
 
+from unittest import skipUnless
+
+import django
 import pytest
 from django import forms
 from django.core import exceptions
 from django.test import SimpleTestCase
+from django.utils.html import escape
 
 from django_mysql.forms import JSONField, SimpleListField, SimpleSetField
 
@@ -306,4 +310,52 @@ class TestJSONField(SimpleTestCase):
     def test_prepare_value(self):
         field = JSONField()
         assert field.prepare_value({'a': 'b'}) == '{"a": "b"}'
+        assert field.prepare_value(['a', 'b']) == '["a", "b"]'
+        assert field.prepare_value(True) == 'true'
+        assert field.prepare_value(False) == 'false'
+        assert field.prepare_value(3.14) == '3.14'
         assert field.prepare_value(None) == 'null'
+        assert field.prepare_value('foo') == '"foo"'
+
+    def test_redisplay_wrong_input(self):
+        """
+        When displaying a bound form (typically due to invalid input), the form
+        should not overquote JSONField inputs.
+        """
+        class JsonForm(forms.Form):
+            name = forms.CharField(max_length=2)
+            jfield = JSONField()
+
+        # JSONField input is fine, name is too long
+        form = JsonForm({'name': 'xyz', 'jfield': '["foo"]'})
+        assert '[&quot;foo&quot;]</textarea>' in form.as_p()
+
+        # This time, the JSONField input is wrong
+        form = JsonForm({'name': 'xy', 'jfield': '{"foo"}'})
+        # Appears once in the textarea and once in the error message
+        assert form.as_p().count(escape('{"foo"}')) == 2
+
+    def test_already_converted_value(self):
+        field = JSONField(required=False)
+        tests = [
+            '["a", "b", "c"]',
+            '{"a": 1, "b": 2}',
+            '1',
+            '1.5',
+            '"foo"',
+            'true',
+            'false',
+            'null',
+        ]
+        for json_string in tests:
+            val = field.clean(json_string)
+            assert field.clean(val) == val
+
+    @skipUnless(django.VERSION[:2] >= (1, 9),
+                "Requires Django 1.9+ for Field.disabled")
+    def test_disabled(self):
+        class JsonForm(forms.Form):
+            jfield = JSONField(disabled=True)
+
+        form = JsonForm({'jfield': '["bar"]'}, initial={'jfield': ['foo']})
+        assert '[&quot;foo&quot;]</textarea>' in form.as_p()