+++ tests/testapp/test_models.py
@@ -7,6 +7,7 @@ import pickle
 import re
 from unittest import mock, skipUnless
 
+import django
 import pytest
 from django.contrib.auth.models import User
 from django.db.models.query import QuerySet
@@ -359,6 +360,12 @@ class FoundRowsTests(TestCase):
         authors = Author.objects.sql_calc_found_rows()[:5]
         with self.assertRaises(RuntimeError):
             authors.found_rows
+
+    @skipUnless(django.VERSION[:2] >= (1, 11), "Need Django 1.11+")
+    def test_it_doesnt_work_with_iterator(self):
+        authors = Author.objects.sql_calc_found_rows()[:5]
+        with pytest.raises(ValueError, message="doesn't work with iterator()"):
+            authors.iterator()
 
     def test_it_working(self):
         with self.assertNumQueries(2):